refactor: toastとsnackbarの共通実装を切り出し
