+++ packages/react/src/components/Notification/layout.css
@@ -1,0 +1,53 @@
+.charcoal-snackbar-region,
+.charcoal-toast-region {
+  position: fixed;
+  left: 0;
+  right: 0;
+  display: grid;
+  justify-content: center;
+  pointer-events: none;
+  box-sizing: border-box;
+}
+
+.charcoal-snackbar,
+.charcoal-toast {
+  box-sizing: border-box;
+  border-radius: var(--charcoal-radius-oval);
+  width: fit-content;
+  max-width: calc(100vw - var(--charcoal-space-46));
+  padding: var(--charcoal-space-25) var(--charcoal-space-40);
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  gap: var(--charcoal-space-20);
+  pointer-events: auto;
+}
+
+.charcoal-snackbar-content,
+.charcoal-toast-content {
+  display: flex;
+  align-items: center;
+  min-width: 0;
+}
+
+.charcoal-snackbar-label,
+.charcoal-toast-label {
+  min-width: 0;
+  overflow: hidden;
+  overflow-wrap: break-word;
+  display: -webkit-box;
+  -webkit-box-orient: vertical;
+  -webkit-line-clamp: 2;
+  font-size: var(--charcoal-text-font-size-caption-m);
+  font-weight: var(--charcoal-text-font-weight-bold);
+  line-height: var(--charcoal-text-line-height-caption-m);
+}
+
+@media (prefers-reduced-motion: reduce) {
+  .charcoal-snackbar,
+  .charcoal-snackbar[data-exiting='true'],
+  .charcoal-toast,
+  .charcoal-toast[data-exiting='true'] {
+    animation: none;
+  }
+}
